some changes in cpccTrustedPlatform

+++ io.cpccTrustedPlatform.h
@@ -88,7 +88,7 @@ public:
         /* Obtain current time as seconds elapsed since the Epoch. */
         current_time = time(NULL);
         firstrun_time=fs.getFileModificationDate(getStorageFolder().c_str());
-        days = (current_time - firstrun_time)/ (60*60*24); // convert from seconds to days
+        days = (int) ((current_time - firstrun_time)/ (60*60*24)); // convert from seconds to days
         
         return days;
     }
